figma-transformer: extract effect overflow policy
