+++ figma-transformer/src/Html/StaticHtmlEmitter.php
@@ -57,6 +57,8 @@ final class StaticHtmlEmitter
 
     private ?TransformDiagnosticsBuilder $transformDiagnosticsBuilder = null;
 
+    private ?EffectOverflowPolicy $effectOverflowPolicy = null;
+
     private function designSystemExtractor(): DesignSystemExtractor
     {
         return $this->designSystemExtractor ??= new DesignSystemExtractor();
@@ -87,6 +89,11 @@ final class StaticHtmlEmitter
     private function transformDiagnosticsBuilder(): TransformDiagnosticsBuilder
     {
         return $this->transformDiagnosticsBuilder ??= new TransformDiagnosticsBuilder();
+    }
+
+    private function effectOverflowPolicy(): EffectOverflowPolicy
+    {
+        return $this->effectOverflowPolicy ??= new EffectOverflowPolicy();
     }
 
     private function layoutIntentClassifier(): LayoutIntentClassifier
@@ -3842,7 +3849,7 @@ final class StaticHtmlEmitter
             }
         }
 
-        if ( true === ($layout['clips_content'] ?? false) && ! $this->containsStickyPrimary($node) && ! $this->clipsVisibleEffectOverflow($node) ) {
+        if ( $this->effectOverflowPolicy()->shouldHideOverflow($node, $this->containsStickyPrimary($node)) ) {
             $styles[] = 'overflow:hidden';
         }
 
@@ -5246,63 +5253,6 @@ final class StaticHtmlEmitter
     private function effectStyles(array $node, string $type): array
     {
         return $this->styleDeclarationBuilder()->effectStyles($node, $type);
-    }
-
-    /**
-     * @param array<string, mixed> $node
-     */
-    private function clipsVisibleEffectOverflow(array $node): bool
-    {
-        $box = is_array($node['box'] ?? null) ? $node['box'] : array();
-        foreach ( $this->nodeList($node) as $child ) {
-            if ( ! is_array($child) ) {
-                continue;
-            }
-            if ( $this->nodeEffectOverflowAmount($child) <= 0.0 ) {
-                if ( $this->clipsVisibleEffectOverflow($child) ) {
-                    return true;
-                }
-                continue;
-            }
-
-            $childBox = is_array($child['box'] ?? null) ? $child['box'] : array();
-            $left = isset($childBox['x']) && is_numeric($childBox['x']) ? (float) $childBox['x'] : 0.0;
-            $top = isset($childBox['y']) && is_numeric($childBox['y']) ? (float) $childBox['y'] : 0.0;
-            $width = isset($childBox['width']) && is_numeric($childBox['width']) ? (float) $childBox['width'] : 0.0;
-            $height = isset($childBox['height']) && is_numeric($childBox['height']) ? (float) $childBox['height'] : 0.0;
-            $parentWidth = isset($box['width']) && is_numeric($box['width']) ? (float) $box['width'] : null;
-            $parentHeight = isset($box['height']) && is_numeric($box['height']) ? (float) $box['height'] : null;
-
-            if ( $left <= 0.0 || $top <= 0.0 || (null !== $parentWidth && $left + $width >= $parentWidth) || (null !== $parentHeight && $top + $height >= $parentHeight) ) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * @param array<string, mixed> $node
-     */
-    private function nodeEffectOverflowAmount(array $node): float
-    {
-        $amount = 0.0;
-        foreach ( is_array($node['figma_effects'] ?? null) ? $node['figma_effects'] : array() as $effect ) {
-            if ( ! is_array($effect) || false === ($effect['visible'] ?? true) ) {
-                continue;
-            }
-            $type = (string) ($effect['type'] ?? '');
-            if ( ! in_array($type, array('drop_shadow', 'layer_blur'), true) ) {
-                continue;
-            }
-            $radius = isset($effect['radius']) && is_numeric($effect['radius']) ? (float) $effect['radius'] : 0.0;
-            $spread = isset($effect['spread']) && is_numeric($effect['spread']) ? (float) $effect['spread'] : 0.0;
-            $offsetX = isset($effect['offset_x']) && is_numeric($effect['offset_x']) ? abs((float) $effect['offset_x']) : 0.0;
-            $offsetY = isset($effect['offset_y']) && is_numeric($effect['offset_y']) ? abs((float) $effect['offset_y']) : 0.0;
-            $amount = max($amount, $radius + $spread + $offsetX, $radius + $spread + $offsetY);
-        }
-
-        return $amount;
     }
 
     /**
